+++ common/src/main/java/io/github/flemmli97/improvedmobs/mixinhelper/INodeBreakable.java
@@ -1,0 +1,12 @@
+package io.github.flemmli97.improvedmobs.mixinhelper;
+
+public interface INodeBreakable {
+
+    void setCanBreakBlocks(boolean flag);
+
+    boolean canBreakBlocks();
+
+    void setCanClimbLadder(boolean flag);
+
+    boolean canClimbLadder();
+}